Configure dynamic options in post_install.php
This simplifies WCFSetup and more importantly PackageInstallationDispatcher. It
also removes the useless session variable for the Imagick default status.
